+++ drizzled/my_decimal.h
@@ -398,7 +398,7 @@ int my_decimal_intg(const my_decimal *a)
 }
 
 
-void my_decimal_trim(ulong *precision, uint *scale);
+void my_decimal_trim(uint32_t *precision, uint *scale);
 
 
 #endif /*my_decimal_h*/
